+++ index/src/assets/mod.rs
@@ -1,0 +1,2 @@
+mod acala;
+pub use acala::*;
