+++ index/src/executors/dex_executors/acala_dot_swap.rs
@@ -1,8 +1,8 @@
 use pink_subrpc::{create_transaction, send_transaction, ExtraParam};
 
-use crate::constants::assets::{CurrencyId, TokenSymbol};
+use crate::assets::{AggregatedSwapPath, CurrencyId, TokenSymbol};
+use crate::prelude::DexExecutor;
 use crate::traits::common::Error;
-use crate::{constants::assets::AggregatedSwapPath, prelude::DexExecutor};
 use alloc::{
     string::{String, ToString},
     vec,
