+++ index/src/lib.rs
@@ -2,6 +2,7 @@
 
 extern crate alloc;
 
+pub mod assets;
 mod constants;
 pub mod executors;
 pub mod graph;
